Time: 6 ms (21.27%), Space: 7.9 MB (5.33%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -1,15 +1,15 @@
 class Solution {
 public:
-    int reverse(int x) {
-        int n=x;
-        vector<int> nums;
+    long long reverse(long long x) {
+        long long n=x;
+        vector<long long> nums;
         while(n){
             nums.push_back(n%10);
             n=n/10;
         }
-        int c=nums.size()-1;
-        int num=0;
-        for(int i: nums){
+        long long c=nums.size()-1;
+        long long num=0;
+        for(long long i: nums){
             num+=i*pow(10,c);
             c--;
         }
